Moved interesting values/defs to common.go - executable now uses Version and Nbt2JsonUrl from nbt2json package
