+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2017 Jim Nelson
+Copyright (c) 2020 Jim Nelson
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
